+++ screens/profileChores/WeekChoresScreen.tsx
@@ -1,10 +1,10 @@
 import { CompositeScreenProps } from '@react-navigation/native';
 import React from 'react';
-import RenderChores from '../component/choreComponent/RenderChores';
-import { ChoreTabScreenProps } from '../navigation/ChoresTabNavigator';
-import { RootStackScreenProps } from '../navigation/RootStackNavigator';
-import { getAllHouseholdsByUserIdSelector } from '../store/household/householdSelectors';
-import { useAppSelector } from '../store/store';
+import RenderChores from '../../component/choreComponent/RenderChores';
+import { ChoreTabScreenProps } from '../../navigation/ChoresTabNavigator';
+import { RootStackScreenProps } from '../../navigation/RootStackNavigator';
+import { getAllHouseholdsByUserIdSelector } from '../../store/household/householdSelectors';
+import { useAppSelector } from '../../store/store';
 
 type Props = CompositeScreenProps<ChoreTabScreenProps<'Week'>, RootStackScreenProps>;
 
@@ -15,9 +15,7 @@ export default function WeekChoresScreen({ navigation, route }: Props) {
 
     const userHousehold = useAppSelector((state) => getAllHouseholdsByUserIdSelector(state, user?.id));
     // console.log(route.params);
-    const currentHousehold = useAppSelector((state) =>
-        state.household.householdList.find((h) => h.id === route.params.householdId)
-    );
+    const currentHousehold = useAppSelector((state) => state.household.householdList.find((h) => h.id === route.params.householdId));
     const members = useAppSelector((state) => state.member.memberList);
     return (
         <RenderChores
